+++ RELEASE_NOTES.md
@@ -32,4 +32,5 @@
 
 ## Version 1.6.5
 - Remove atomic operations on non atomic bool in scheduler
+- Support removed for Debian 11