+++ RELEASE_NOTES.md
@@ -32,8 +32,7 @@
 
 ## Version 1.6.5
 - Remove atomic operations on non atomic bool in scheduler
-- Support removed for Debian 11, Ubuntu 20.04 and Fedora 42
-- Support added for Fedora 44
-- Support added for Alpine 2.24, removed for 3.20 and 3.21
+- Support removed for Debian 11, Ubuntu 20.04, Fedora 42 Alpine 3.20 and Alpine 3.21
+- Support added for Fedora 44 and Alpine 3.24
 - Added function `iot_data_vector_to_list` to populate a list from a vector